tests/wpt/tests/html/canvas/offscreen/line-styles/2d.line.miter.lineedge.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.line.miter.lineedge</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.line.miter.lineedge</h1>
<p class="desc">Miter joins are not drawn when the miter limit is exceeded at the corners of a zero-height rectangle</p>


<script>
var t = async_test("Miter joins are not drawn when the miter limit is exceeded at the corners of a zero-height rectangle");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);

  ctx.lineWidth = 200;
  ctx.lineJoin = 'miter';

  ctx.strokeStyle = '#f00';
  ctx.miterLimit = 1.414;
  ctx.beginPath();
  ctx.strokeRect(100, 25, 200, 0);

  _assertPixel(canvas, 1,1, 0,255,0,255);
  _assertPixel(canvas, 48,1, 0,255,0,255);
  _assertPixel(canvas, 48,48, 0,255,0,255);
  _assertPixel(canvas, 1,48, 0,255,0,255);
  t.done();

});
</script>
